requirements.txt
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
mlflow>=2.8.0
sentencepiece>=0.1.99
protobuf>=3.20.0
numpy>=1.24.0
packaging>=21.0
safetensors>=0.4.0
tensorboard>=2.14.0
psutil>=5.9.0
pyyaml>=6.0
tqdm>=4.66.0
deepspeed>=0.12.0

